pyproject.toml
[project]
name = "bulk-migrator"
version = "2.3.2"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "msal>=1.33.0",
    "pytest>=8.4.1",
    "pytest-cov>=6.2.1",
    "pytest-mock>=3.14.1",
    "pytest-xdist>=3.6.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "ruff>=0.8.0",
    "mypy>=1.8.0",
    "types-requests>=2.32.0",
    "pre-commit>=4.0.0",
    "bandit[toml]>=1.7.5",
    "pip-audit>=2.6.1",
    "cyclonedx-bom>=4.0.0",
]

[tool.ruff]
line-length = 88
target-version = "py313"
exclude = ["logs/", ".venv/", "build/", ".mypy_cache/", ".pytest_cache/", ".ruff_cache/"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "C90", "T20"]

[tool.ruff.lint.isort]
known-first-party = ["src"]

[tool.mypy]
python_version = "3.13"
warn_unused_ignores = true
warn_redundant_casts = true
no_implicit_optional = true
# 段階的厳格化: staging で追加設定を有効化
# disallow_any_generics = true  # staging で有効化予定
# warn_return_any = true        # staging で有効化予定
# strict_equality = true        # staging で有効化予定

# ローカルモジュールの import エラーを無視
[[tool.mypy.overrides]]
module = [
    "config_manager",
    "logger",
    "structured_logger",
    "transfer",
    "skiplist",
    "rebuild_skip_list",
    "src.*",
    "msal"
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests/unit", "tests/integration", "tests/security"]
python_files = ["test_*.py"]
addopts = [
    "--junit-xml=pytest-results.xml",
    "-v"
]
markers = [
    "unit: 単体テスト",
    "integration: 統合テスト",
    "auth: 認証関連テスト",
    "transfer: ファイル転送テスト",
    "config: 設定関連テスト"
]
[tool.bandit]
exclude_dirs = ["tests", ".venv", "build"]
skips = ["B101"]  # assert_used - テストで使用するため除外

[tool.bandit.assert_used]
skips = ["**/test_*.py", "**/tests/**"]
